+++ src/components/NcHeaderMenu/NcHeaderMenu.vue
@@ -60,25 +60,22 @@ export default {
 	<component :is="wrapperTag"
 		:id="id"
 		ref="headerMenu"
-		v-click-outside="clickOutsideConfig"
 		:aria-labelledby="isNav ? triggerId : null"
-		:class="{ 'header-menu--opened': opened }"
+		:class="{ 'header-menu--opened': isOpened }"
 		class="header-menu"
-		v-on="listeners">
+		@focusout="onFocusOut">
 		<!-- Trigger -->
 		<NcButton :id="isNav ? triggerId : null"
-			ref="trigger"
+			ref="triggerButton"
 			type="tertiary-no-background"
 			class="header-menu__trigger"
 			:aria-label="ariaLabel"
 			:aria-describedby="description ? descriptionId : null"
 			:aria-controls="`header-menu-${id}`"
-			:aria-expanded="opened.toString()"
+			:aria-expanded="isOpened.toString()"
 			size="large"
 			@click.prevent="toggleMenu">
 			<template #icon>
-				<!-- @slot Icon trigger slot. Make sure the svg path
-				is at least 16px. Usually mdi icon works at 20px -->
 				<slot name="trigger" />
 			</template>
 		</NcButton>
@@ -90,244 +87,197 @@ export default {
 		</span>
 
 		<!-- Visual triangle -->
-		<div v-show="opened" class="header-menu__carret" />
+		<div v-show="isOpened" class="header-menu__caret" />
 
 		<!-- Menu opened content -->
-		<div v-show="opened"
+		<div v-show="isOpened"
 			:id="`header-menu-${id}`"
 			class="header-menu__wrapper">
-			<div ref="content" class="header-menu__content">
-				<!-- @slot Main content -->
+			<div ref="contentContainer" class="header-menu__content">
 				<slot />
 			</div>
 		</div>
 	</component>
 </template>
 
-<script>
-import { vOnClickOutside as ClickOutside } from '@vueuse/components'
+<script setup lang="ts">
+import type { FocusTrap } from 'focus-trap'
+
+import { onClickOutside } from '@vueuse/core'
 import { createFocusTrap } from 'focus-trap'
+import { computed, nextTick, ref, useTemplateRef, watch, type Slot } from 'vue'
+import { getTrapStack } from '../../utils/focusTrap.js'
+import { useHotKey } from '../../composables/index.js'
 
 import GenRandomId from '../../utils/GenRandomId.js'
-import { clickOutsideOptions } from '../../mixins/index.js'
-import { getTrapStack } from '../../utils/focusTrap.js'
-
 import NcButton from '../NcButton/index.ts'
 
-export default {
-	name: 'NcHeaderMenu',
+const {
+	excludeClickOutsideSelectors = [],
+	open = false,
+	isNav = false,
+} = defineProps<{
+	/**
+	 * Unique id for this menu
+	 */
+	id: string
 
-	components: {
-		NcButton,
-	},
+	/**
+	 * aria-label attribute of the menu open button
+	 */
+	ariaLabel?: string
 
-	directives: {
-		ClickOutside,
-	},
+	/**
+	 * Current menu open state
+	 */
+	open?: boolean
 
-	mixins: [
-		clickOutsideOptions,
-	],
+	/**
+	 * Pass `true` if the header menu is used for website navigation
+	 *
+	 * The wrapper tag will be set to `nav` and its `aria-labelledby`
+	 * will be associated with the menu open button
+	 */
+	isNav?: boolean
 
-	props: {
-		/**
-		 * Unique id for this menu
-		 */
-		id: {
-			type: String,
-			required: true,
-		},
+	/**
+	 * Additional visually hidden description text for the menu
+	 * open button
+	 */
+	description?: string
 
-		/**
-		 * aria-label attribute of the menu open button
-		 */
-		ariaLabel: {
-			type: String,
-			default: '',
-		},
+	/**
+	 * A query-selector or an array of query-selectors
+	 * to be ignored when clicking outside an element
+	 */
+	excludeClickOutsideSelectors?: string | string[]
+}>()
 
-		/**
-		 * Current menu open state
-		 */
-		open: {
-			type: Boolean,
-			default: false,
-		},
+const emit = defineEmits<{
+	/** Emitted when the menu is fully closed (animation done) */
+	closed: []
 
-		/**
-		 * Pass `true` if the header menu is used for website navigation
-		 *
-		 * The wrapper tag will be set to `nav` and its `aria-labelledby`
-		 * will be associated with the menu open button
-		 */
-		isNav: {
-			type: Boolean,
-			default: false,
-		},
+	/** Emitted when the menu is fully opened (animation done) */
+	opened: []
 
-		/**
-		 * Additional visually hidden description text for the menu
-		 * open button
-		 */
-		description: {
-			type: String,
-			default: null,
-		},
-	},
+	/** Updated open state */
+	'update:open': [v: boolean]
+}>()
 
-	emits: [
-		'close',
-		'closed',
-		'open',
-		'opened',
-		'update:open',
-		'cancel',
-	],
+defineSlots<{
+	/** The menu content */
+	default: Slot
+	/** Icon trigger slot. Make sure the svg path is at least 16px. Usually mdi icon works at 20px */
+	trigger: Slot
+}>()
 
-	data() {
-		return {
-			focusTrap: null,
-			opened: this.open,
-			shortcutsDisabled: window.OCP?.Accessibility?.disableKeyboardShortcuts?.(),
-			triggerId: GenRandomId(),
-			descriptionId: GenRandomId(),
-		}
-	},
+/** Id of the menu description */
+const descriptionId = GenRandomId()
+/** Id of the trigger button */
+const triggerId = GenRandomId()
+/** The active focus trap (if any) */
+const focusTrap = ref<FocusTrap>()
+/** Is the menu currently opened */
+const isOpened = ref(open)
+/** HTML tag to use for the header menu */
+const wrapperTag = computed(() => isNav ? 'nav' : 'div')
 
-	computed: {
-		wrapperTag() {
-			return this.isNav ? 'nav' : 'div'
-		},
+/** The menu content container element */
+const contentContainer = useTemplateRef('contentContainer')
+/** The overall header menu wrapping element (<nav> or <div>) */
+const headerMenu = useTemplateRef<HTMLElement>('headerMenu')
+/** The menu trigger button */
+const triggerButton = useTemplateRef('triggerButton')
 
-		clickOutsideConfig() {
-			return [
-				this.closeMenu,
-				this.clickOutsideOptions,
-			]
-		},
+// Handle click outside of the menu -> should close the menu
+const ignore = computed(() => Array.isArray(excludeClickOutsideSelectors)
+	? excludeClickOutsideSelectors
+	: excludeClickOutsideSelectors.split(' '),
+)
+onClickOutside(headerMenu, () => setMenuState(false), { ignore })
 
-		listeners() {
-			if (this.isNav) {
-				return {
-					focusout: this.onFocusOut,
-				}
-			}
-			return null
-		},
-	},
+// Pressing escape should close the menu
+useHotKey('Escape', () => setMenuState(false), { prevent: true })
 
-	watch: {
-		open(open) {
-			if (open) {
-				this.openMenu()
-			} else {
-				this.closeMenu()
-			}
-		},
-	},
+// Watch the open prop to adjust the internal opened state
+watch(() => open, (state: boolean) => setMenuState(state))
 
-	mounted() {
-		document.addEventListener('keydown', this.onKeyDown)
-	},
-	beforeUnmount() {
-		document.removeEventListener('keydown', this.onKeyDown)
-	},
+/**
+ * Toggle the current menu open state
+ */
+function toggleMenu() {
+	setMenuState(!isOpened.value)
+}
 
-	methods: {
-		/**
-		 * Toggle the current menu open state
-		 */
-		toggleMenu() {
-			// Toggling current state
-			if (!this.opened) {
-				this.openMenu()
-			} else {
-				this.closeMenu()
-			}
-		},
+/**
+ * Set the menu opened state
+ * @param state The opened state to set
+ */
+async function setMenuState(state: boolean) {
+	if (state === isOpened.value) {
+		return
+	}
 
-		/**
-		 * Close the current menu
-		 *
-		 * @param {boolean} cancelled emit a cancel event instead of close
-		 */
-		closeMenu(cancelled = false) {
-			// Close the menu
-			this.opened = false
-			this.$emit(cancelled ? 'cancel' : 'close')
-			this.$emit('update:open', false)
+	isOpened.value = state
+	emit('update:open', state)
 
-			// Kill focus trap
-			this.clearFocusTrap()
+	// wait one tick to make sure the rendering finished
+	await nextTick()
+	// either add or clear the focus trap
+	await (state ? addFocusTrap() : clearFocusTrap())
 
-			// Wait for component to finish rendering
-			this.$nextTick(() => {
-				this.$emit('closed')
-			})
-		},
+	// Emit signal to mark finished toggling
+	// @ts-expect-error This seems to be broken in Vue's typescript macro compiler...
+	emit(state ? 'opened' : 'closed')
+}
 
-		/**
-		 * Open the current menu
-		 */
-		openMenu() {
-			// Open the menu
-			this.opened = true
-			this.$emit('open')
-			this.$emit('update:open', true)
+/**
+ * When this is role navigation, then we cannot apply a focus trap.
+ * In this case we close the menu on focus-out.
+ */
+function onFocusOut(event: FocusEvent) {
+	// Is not a navigation
+	if (!isNav) {
+		return
+	}
 
-			// Wait for component to finish rendering
-			this.$nextTick(() => {
-				this.useFocusTrap()
-				this.$emit('opened')
-			})
-		},
+	// Event target is not a node
+	if (!(event.relatedTarget instanceof Node)) {
+		return
+	}
 
-		onKeyDown(event) {
-			if (this.shortcutsDisabled || !this.opened) {
-				return
-			}
+	if (headerMenu.value?.contains(event.relatedTarget)) {
+		setMenuState(false)
+	}
+}
 
-			// If escape have been pressed, we close
-			if (event.key === 'Escape') {
-				event.preventDefault()
+/**
+ * Add focus trap for accessibility.
+ * Shall only be used when all children are mounted
+ * and available in the DOM. We use $nextTick for that.
+ */
+async function addFocusTrap() {
+	// We cannot add the focus trap on navigation roles
+	// also skip if already set
+	if (isNav || focusTrap.value) {
+		return
+	}
 
-				/** User cancelled the menu by pressing escape */
-				this.closeMenu(true)
-			}
-		},
+	// Init focus trap
+	focusTrap.value = createFocusTrap(contentContainer.value!, {
+		allowOutsideClick: true,
+		trapStack: getTrapStack(),
+		fallbackFocus: triggerButton.value?.$el,
+	})
+	focusTrap.value.activate()
+}
 
-		/**
-		 * @param {FocusEvent} event The focus event
-		 */
-		onFocusOut(event) {
-			if (!this.$refs.headerMenu.contains(event.relatedTarget)) {
-				this.closeMenu()
-			}
-		},
-
-		/**
-		 * Add focus trap for accessibility.
-		 * Shall only be used when all children are mounted
-		 * and available in the DOM. We use $nextTick for that.
-		 */
-		async useFocusTrap() {
-			if (this.isNav || this.focusTrap) {
-				return
-			}
-			// Init focus trap
-			const contentContainer = this.$refs.content
-			this.focusTrap = createFocusTrap(contentContainer, {
-				allowOutsideClick: true,
-				trapStack: getTrapStack(),
-				fallbackFocus: this.$refs.trigger,
-			})
-			this.focusTrap.activate()
-		},
-		clearFocusTrap() {
-			this.focusTrap?.deactivate()
-			this.focusTrap = null
-		},
-	},
+/**
+ * Deactivate and clear the focus trap
+ */
+function clearFocusTrap() {
+	focusTrap.value?.deactivate()
+	focusTrap.value = undefined
 }
 </script>
 
@@ -336,7 +286,7 @@ export default {
 // Also used for menu top-right positioning
 $externalMargin: 8px;
 
-@import './header-menu__trigger';
+@use './header-menu__trigger' as *;
 
 .header-menu {
 	&__wrapper {